requirements.txt
asgiref==3.6.0
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
whitenoise==6.4.0
